requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
tzdata==2022.1
Pillow==9.2.0
djangorestframework==3.13.1